Further fixing of warning produced by minority team members.

+++ COMG_DonovanSim/Source/SP2.h
@@ -619,7 +619,7 @@ private:
 	bool wingsFound;
 	bool engineFound;
 	//CRYSTAL RELATED STUFF   
-	bool checkCrystalPos(int xcoord, int zcoord, int i);
+	bool checkCrystalPos(float xcoord, float zcoord, int i);
 
 	void crystalgen();
 	void asteroidgen();
